# api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Alba API
  version: "1.0.0"
  description: >
    Wire contract for the Alba credit pre-qualification demo.
    ARCHITECTURE.md decides behavior. This file decides the JSON.
    api/contract_models.py and web/src/api/schema.d.ts are generated from this file.
    Do not hand-write a second request or response type.
servers:
  - url: http://localhost:8000
    description: API process. The web app calls the same paths under the /api proxy, which strips that prefix.
...
security:
  - bearerAuth: []
paths:
...
  /consultants/search:
    get:
      tags: [Session]
      operationId: searchConsultants
      security: []
      summary: Demo helper. Search the active consultants, or pick one at random.
      description: >
        Only with DEMO_LOGIN=1; not found otherwise. It fills the consultant login form and never opens a session.
        Send q or random=true, not both and not neither.
        Only consultants whose service_agents.agent_status is Active are listed, because only they can receive a code.
        With q, at most 20 rows matching name, employee code, or consultant_id,
        ordered by last_name, first_name, consultant_id ascending.
        With random=true, one active consultant.
      parameters:
        - $ref: "#/components/parameters/OptionalSearchQuery"
        - $ref: "#/components/parameters/RandomPick"
      responses:
        "200":
          description: Matching active consultants.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultantSearchResults"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/InvalidRequest"
...
  /consultant/me:
    get:
      tags: [Consultant]
      operationId: getCurrentConsultant
      summary: The session consultant's id, employee code, name, and specialty.
      description: >
        Requires role consultant. consultant_id is the token sub. specialty is null when the file has none.
        The consultant screens read the name from here; the login responses never carry it.
      responses:
        "200":
          description: The consultant on the token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentConsultant"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OptionalSearchQuery:
      name: q
      in: query
      required: false
      schema:
        type: string
        minLength: 1
    RandomPick:
      name: random
      in: query
      required: false
      schema:
        type: boolean
        enum: [true]
...
  responses:
    Unauthorized:
      description: Missing token, expired token, or a code that does not open a session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UnauthorizedError"
    Forbidden:
      description: The token role cannot call this route.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ForbiddenError"
    NotFound:
      description: No row for this id under the token's scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotFoundError"
...
    InvalidRequest:
      description: The body or query does not match the schema.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InvalidRequestError"
  schemas:
...
    UnauthorizedError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          enum: [unauthorized]
    ForbiddenError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          enum: [forbidden]
    NotFoundError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          enum: [not_found]
...
    InvalidRequestError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          enum: [invalid_body]
...
    ConsultantSearchHit:
      type: object
      additionalProperties: false
      required: [consultant_id, employee_code, first_name, last_name, email]
      properties:
        consultant_id:
          type: string
        employee_code:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
    ConsultantSearchResults:
      type: object
      additionalProperties: false
      required: [consultants]
      properties:
        consultants:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/ConsultantSearchHit"
...
    CurrentConsultant:
      type: object
      additionalProperties: false
      required: [consultant_id, employee_code, first_name, last_name, specialty]
      properties:
        consultant_id:
          type: string
        employee_code:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        specialty:
          type: string
          nullable: true
...